Refactor: /race Directory Structure
-move components into components folder
-create components/cards
-rename card components as cards
-create components/race
-create race component for each page
-move race-helpers.ts into /components/race as utils.ts
-remove unused route /multiplayer/[id]
-create /friends page
-consolidate loader functions
