--- packages/app/src/app/race/(play)/multiplayer/[roomId]/page.tsx
@@ -1,3 +1,0 @@
-export default function PrivateRacePage() {
-  return <div>Private room page not implemented</div>;
-}
