--- packages/app/src/app/race/(play)/practice/loaders.ts
@@ -1,9 +1,0 @@
-import { prisma } from "@/lib/prisma";
-
-export async function getSnippetById(snippetId: string) {
-  return await prisma.snippet.findUnique({
-    where: {
-      id: snippetId,
-    },
-  });
-}
