+++ packages/app/src/app/race/_components/cards/friends-race-card.tsx
@@ -5,9 +5,9 @@ import { Users } from "lucide-react";
 import { useSession } from "next-auth/react";
 import { useToast } from "@/components/ui/use-toast";
 import { bruno_ace_sc } from "@/lib/fonts";
-import { createPrivateRaceRoom } from "../actions";
+import { createPrivateRaceRoom } from "../../actions";
 
-export default function FriendsRace({ enabled }: { enabled: boolean }) {
+export default function FriendsRaceCard({ enabled }: { enabled: boolean }) {
   const { data: session } = useSession();
   const { toast } = useToast();
