+++ packages/app/src/app/race/_components/race/race-details.tsx
